Remove count from onBack lambdas
NavDisplay and SceneStrategy previously
passed an onBack: (count: Int) -> Unit lambda,
which forced every developer implementing that
lambda to take into account the rare cases where
multiple things were popped at once, despite the
common case being only to pop a single entry
at a time.

It turns out: most developers would forget to handle
this case. Instead, by removing the count parameter
from the lambda entirely, the whole system becomes
easier to reason about and the few places where
popping multiple things at a time may happen, we
can call the lambda multiple times.

Relnote: "The `onBack` lambda passed into `NavDisplay` and
to your `SceneStrategy` no longer provides a `count: Int`
parameter to indicate when multiple entries should be popped.
Instead, the `() -> Unit` lambda will now be called multiple
times in a row in the rare case where your Scene requests
popping multiple entries."
Test: existing tests all pass
BUG: 446989346

+++ navigation3/navigation3-ui/src/commonMain/kotlin/androidx/navigation3/scene/SceneState.kt
@@ -30,14 +30,13 @@ import kotlin.collections.plusAssign
  *
  * @param entries all of the entries that are associated with this state
  * @param sceneStrategy the [SceneStrategy] to determine which scene to render a list of entries.
- * @param onBack a callback for handling system back press. The passed [Int] refers to the number of
- *   entries to pop from the end of the backstack, as calculated by the [sceneStrategy].
+ * @param onBack a callback for handling system back press.
  */
 @Composable
 public fun <T : Any> rememberSceneState(
     entries: List<NavEntry<T>>,
     sceneStrategy: SceneStrategy<T>,
-    onBack: (Int) -> Unit,
+    onBack: () -> Unit,
 ): SceneState<T> {
     // re-wrap the entries with the SceneSetupNavEntryDecorator to ensure all the ensures are
     // inside of a moveable content.
